Growth + Start with reprod
Included a simple growth function, each individual grows with a random
factor between 1 and 1.2. This can probably be made more realistic in
the future…

Besides, started with the reproduction. So far I only wrote the part to
retrieve the indices of alive males and alive females. Next up is using
these indices to create the desired interactions (should not be too
difficult, I guess eeeeh… hope..)